file_agent/requirements.txt
pydantic>=2.0.0
pydantic-ai>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
colorlog>=6.0.0
httpx>=0.24.0
streamlit>=1.30.0
@modelcontextprotocol/server-filesystem
